rename get user stats submit
